[bug 859973] Improve groups listing view
This commit hardcodes ordering of groups alphabetically. It creates a
list structure for displaying groups. The query is filtered using a
distinct clause

A new pagination variable has been introduced in apps.phonebook.forms to
take care of 50 items per page.

A simple icon is used to represent members count, bypassing the use of
pluralize or similar filters. Appearance can, of course, be improved
with CSS, but I didn't touch any CSS files in this commit.